requirements.txt

streamlit
python-dotenv
pytesseract
Pillow
openai
langchain
PyMuPDF
